Trying to decide where to put SyncSkyscraper...
