+++ src/ElevatorSharp.Web/Controllers/ElevatorController.cs
@@ -11,7 +11,7 @@ using Newtonsoft.Json;
 
 namespace ElevatorSharp.Web.Controllers
 {
-    public class ElevatorController : Controller
+    public class ElevatorController : BaseController
     {
         #region Elevator Events
         /// <summary>
@@ -74,58 +74,6 @@ namespace ElevatorSharp.Web.Controllers
         {
             var json = JsonConvert.SerializeObject(elevatorDto);
             return Content(json, "application/json");
-        }
-        #endregion
-
-        #region Helper Methods
-        private static Skyscraper LoadSkyscraper()
-        {
-            var cache = MemoryCache.Default;
-            if (cache.Contains("skyscraper"))
-            {
-                return (Skyscraper)cache.Get("skyscraper");
-            }
-            return null;
-        }
-
-        /// <summary>
-        /// Transfer all required data from client.
-        /// </summary>
-        /// <param name="elevatorDto"></param>
-        /// <returns></returns>
-        private static Skyscraper SyncSkyscraper(ElevatorDto elevatorDto)
-        {
-            var skyscraper = LoadSkyscraper();
-            skyscraper.Elevators[elevatorDto.ElevatorIndex].PressedFloors = elevatorDto.PressedFloors;
-
-            skyscraper.Elevators[elevatorDto.ElevatorIndex].DestinationQueue.Clear();
-            if (elevatorDto.DestinationQueue != null)
-            {
-                foreach (var floor in elevatorDto.DestinationQueue)
-                    skyscraper.Elevators[elevatorDto.ElevatorIndex].DestinationQueue.Enqueue(floor); 
-            }
-            
-            return skyscraper;
-        }
-
-        /// <summary>
-        /// Use ElevatorCommands for sending back to client.
-        /// </summary>
-        /// <param name="elevatorDto"></param>
-        /// <param name="skyscraper"></param>
-        /// <returns></returns>
-        private static ElevatorCommands CreateElevatorCommands(ElevatorDto elevatorDto, Skyscraper skyscraper)
-        {
-            var elevatorCommands = new ElevatorCommands { ElevatorIndex = elevatorDto.ElevatorIndex };
-            var destinationQueue = skyscraper.Elevators[elevatorDto.ElevatorIndex].DestinationQueue;
-            while (destinationQueue.Count > 0)
-            {
-                var destination = destinationQueue.Dequeue();
-
-                // TODO: does not take JumpQueue into account
-                elevatorCommands.GoToFloor.Enqueue(new GoToFloorCommand(destination, false));
-            }
-            return elevatorCommands;
         }
         #endregion
     }
