+++ src/ElevatorSharp.Web/ElevatorSharp.Web.csproj
@@ -168,6 +168,7 @@
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
+    <Compile Include="Controllers\BaseController.cs" />
     <Compile Include="Controllers\ElevatorController.cs" />
     <Compile Include="Controllers\FloorController.cs" />
     <Compile Include="Controllers\SkyscraperController.cs" />
@@ -273,6 +274,7 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
+    <Folder Include="Views\Base\" />
     <Folder Include="Views\Elevator\" />
     <Folder Include="Views\Floor\" />
   </ItemGroup>
